requirements.txt
PyQt5
numpy
opencv-python
tqdm
joblib
scipy
matplotlib
numpy
pascal-voc-tools
pascal-voc-writer
pytest-qt
pyrealsense2
python-pcl
